+++ src/load_balancer.rs
@@ -4197,7 +4197,7 @@ impl LoadBalancer {
             if let Some(ref ps) = h.proxy_passive
                 && let Some(entry) = ps.unhealthy.get(&self.host_port_keys[i])
             {
-                passive_ejected.push((i, entry.ejected_at_ms));
+                passive_ejected.push((i, entry.ejected_at_tick_ms));
                 continue;
             }
             bitset.set(i);
@@ -4253,7 +4253,7 @@ impl LoadBalancer {
             if let Some(ref ps) = h.proxy_passive
                 && let Some(entry) = ps.unhealthy.get(&self.host_port_keys[i])
             {
-                passive_ejected.push((i, entry.ejected_at_ms));
+                passive_ejected.push((i, entry.ejected_at_tick_ms));
                 continue;
             }
             bitset.set(i);
@@ -4311,7 +4311,7 @@ impl LoadBalancer {
             if let Some(ref ps) = h.proxy_passive
                 && let Some(entry) = ps.unhealthy.get(&self.host_port_keys[i])
             {
-                passive_ejected.push((i, entry.ejected_at_ms));
+                passive_ejected.push((i, entry.ejected_at_tick_ms));
                 return;
             }
             bitset.set(i);
@@ -4357,7 +4357,7 @@ impl LoadBalancer {
             if let Some(ref ps) = h.proxy_passive
                 && let Some(entry) = ps.unhealthy.get(&self.host_port_keys[i])
             {
-                passive_ejected.push((i, entry.ejected_at_ms));
+                passive_ejected.push((i, entry.ejected_at_tick_ms));
                 continue;
             }
             healthy.push((i, target));
@@ -4411,7 +4411,7 @@ impl LoadBalancer {
             if let Some(ref ps) = h.proxy_passive
                 && let Some(entry) = ps.unhealthy.get(&self.host_port_keys[i])
             {
-                passive_ejected.push((i, entry.ejected_at_ms));
+                passive_ejected.push((i, entry.ejected_at_tick_ms));
                 continue;
             }
             healthy.push((i, target));